Implements Issue #4: Wear Guidelines
Wear app has been reworked to run as a service with ongoing notification for
control. Compass is launched upon navigation to new geocache, but it can now
be closed (and the screen can be turned off) and restored through a
button on the notification.

This refactoring may have finally fixed some device-specific bugs I
couldn't track down (like Issue #2 with G Watch R) because most of the
Wear code was changed.