Added support for more vars to MessageDataset
Also significantly altered LocationUtils so it can operate in a more
modular fashion (ie. can be used to track just distance OR direction, or
both). This will make implementing compass support on the Wear side
easier, as GPS location will always be tracked on the phone side for this
app.